Added some stuff to the PN532
- buffer reading and checking (need to change)

+++ Code/sensOrb/lib/BNO08x/sh2_util.c
@@ -47,7 +47,7 @@ void writeu16(uint8_t * p, uint16_t value)
 
 uint32_t readu32(const uint8_t *p)
 {
-    uint32_t retval = p[0] | (p[1] << 8) | (p[2] << 16) | (p[3] << 24);
+    uint32_t retval = p[0] | ((uint32_t)p[1] << 8) | ((uint32_t)p[2] << 16) | ((uint32_t)p[3] << 24);
     return retval;
 }
 
@@ -88,7 +88,7 @@ void write16(uint8_t * p, int16_t value)
 
 int32_t read32(const uint8_t *p)
 {
-    int32_t retval = p[0] | (p[1] << 8) | (p[2] << 16) | (p[3] << 24);
+    int32_t retval = p[0] | ((uint32_t)p[1] << 8) | ((uint32_t)p[2] << 16) | ((uint32_t)p[3] << 24);
     return retval;
 }
